+++ Localization/enUS.Plugins.lua
@@ -16,6 +16,14 @@ L["Better Wardrobe Desc"] = [=[This plugin fixes the link to the wardrobe that i
 
 There are no options.]=]
 L["Calendar Button Position Overwrite Desc"] = "The position settings are not active when {addon} is enabled."
+L["EllesmereUI"] = true
+L["EllesmereUI Desc"] = [=[Each of the options below is controlled by EllesmereUI and is purely informational.
+
+To change these, go to Game Menu -> EllesmereUI -> Blizz UI Enhanced -> Blizzard Window Skins -> Third-Party Addons and toggle the desired option.
+]=]
+L["EllesmereUI Skin Achievements"] = "Skin Achievements"
+L["EllesmereUI Skin Achievements Desc"] = [=[Applies the EllesmereUI skin to the Achievements Window.
+-> Blizz UI Enhanced -> Blizzard Window Skins -> Third-Party Addons]=]
 L["ElvUI"] = true
 L["ElvUI Desc"] = [=[Each of the options below is controlled by ElvUI and is purely informational.